# wtw_backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Web Toons Wala API",
    "version": "1.0.0",
    "description": "API documentation for Web Toons Wala application"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "ERROR_CODE"
              },
              "message": {
                "type": "string",
                "example": "Detailed error message"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Success message"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          }
        }
      },
...
      "FileUploadResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "File uploaded successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "originalname": {
                "type": "string",
                "example": "example.jpg"
              },
              "filename": {
                "type": "string",
                "example": "file-1234567890.jpg"
              },
              "fileType": {
                "type": "string",
                "enum": ["IMAGE", "AUDIO"],
                "example": "IMAGE"
              },
              "fileURL": {
                "type": "string",
                "example": "https://example.com/file-1234567890.jpg"
              },
              "length": {
                "type": "number",
                "description": "Audio duration in seconds (only for AUDIO type)",
                "example": 120.5
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
    "/user/login": {
      "post": {
        "tags": ["Users"],
        "summary": "User login",
        "description": "Authenticate a user and return a token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "johndoe"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/getAllUsers": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "description": "Retrieve a list of all users",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/addNewUser": {
      "post": {
        "tags": ["Users"],
        "summary": "Add a new user",
        "description": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                  },
                  "username": {
                    "type": "string",
                    "example": "johndoe"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  },
                  "userType": {
                    "type": "string",
                    "enum": ["ADMIN", "CUSTOMER"],
                    "example": "CUSTOMER"
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "username",
                  "password",
                  "userType"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/logout": {
      "get": {
        "tags": ["Users"],
        "summary": "User logout",
        "description": "Logout the current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/upload/uploadFile": {
      "post": {
        "tags": ["File Upload"],
        "summary": "Upload a file",
        "description": "Upload a file to the server",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The file to upload"
                  },
                  "fileType": {
                    "type": "string",
                    "enum": ["IMAGE", "AUDIO"],
                    "description": "Type of file being uploaded",
                    "example": "IMAGE"
                  }
                },
                "required": ["file", "fileType"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or file not uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/upload/getFileFromCloud": {
      "get": {
        "tags": ["File Upload"],
        "summary": "Get a file from cloud",
        "description": "Retrieve a file from the cloud storage",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "The key of the file in cloud storage",
            "required": true,
            "schema": {
              "type": "string",
              "example": "file-1234567890.jpg"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File retrieved successfully",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or file not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/upload/getPreSignedURL": {
      "get": {
        "tags": ["File Upload"],
        "summary": "Get a pre-signed URL",
        "description": "Get a pre-signed URL for a file in cloud storage",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "The key of the file in cloud storage",
            "required": true,
            "schema": {
              "type": "string",
              "example": "file-1234567890.jpg"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pre-signed URL retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or file not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}